Remove Document::PaginatedHistory
This commit tidies up remnants of Document::PaginatedHistory and the
implementation of split document history and notes which was used in the
Bootstrap layout.

- Relocate Document::PaginatedHistory::AuditTrailEntry into
  Document::PaginatedTimeline. I've taken this opportunity to rename the
  class to VersionPresenter which better represents its purpose as a
  lightweight presenter wrapper around Version objects appearing within
  the Document Timeline.
- Remove unused Edition 'history' endpoint and views.
- Remove the Audit Trail controller which used to respond to Ajax calls
  for pagination.
- Remove Document::PaginatedHistory and associated tests.
