--- app/views/admin/edition_audit_trail/index.html.erb
@@ -1,8 +1,0 @@
-<div class="audit-trail-page">
-  <h1>Activity</h1>
-  <%= paginate @document_history.query, theme: 'audit_trail' %>
-  <ul class="list-unstyled">
-    <%= render partial: "audit_trail_entry", collection: @document_history.audit_trail %>
-  </ul>
-  <%= paginate @document_history.query, theme: 'audit_trail' %>
-</div>
