--- app/views/admin/editions/history.html.erb
@@ -1,19 +1,0 @@
-<% page_title "History" %>
-
-<div class="row">
-  <section class="col-md-8">
-    <span class="back">
-      <%= link_to 'Back', admin_edition_path(@edition) %>
-    </span>
-
-    <div class="audit-trail-page">
-      <h1>History</h1>
-
-      <%= paginate @document_history.query, theme: 'audit_trail' %>
-      <ul class="list-unstyled">
-        <%= render partial: "admin/editions/audit_trail_entry", collection: @document_history.audit_trail %>
-      </ul>
-      <%= paginate @document_history.query, theme: 'audit_trail' %>
-    </div>
-  </section>
-</div>
